mock-fabric-inbox/requirements.txt
jsonify
Flask
PyJWT